net: add ip address to colima list
