+++ environment/vm/lima/lima.go
@@ -1,14 +1,12 @@
 package lima
 
 import (
-	"bufio"
 	"bytes"
 	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/abiosoft/colima/cli"
@@ -407,74 +405,4 @@ func (l limaVM) User() (string, error) {
 func (l limaVM) Arch() environment.Arch {
 	a, _ := l.RunOutput("uname", "-m")
 	return environment.Arch(a)
-}
-
-// InstanceInfo is the information about a Lima instance
-type InstanceInfo struct {
-	Name   string `json:"name,omitempty"`
-	Status string `json:"status,omitempty"`
-	Arch   string `json:"arch,omitempty"`
-	CPU    int    `json:"cpus,omitempty"`
-	Memory int64  `json:"memory,omitempty"`
-	Disk   int64  `json:"disk,omitempty"`
-}
-
-// Instances returns Lima instances created by colima.
-func Instances() ([]InstanceInfo, error) {
-	var buf bytes.Buffer
-	cmd := cli.Command("limactl", "list", "--json")
-	cmd.Stdout = &buf
-
-	if err := cmd.Run(); err != nil {
-		return nil, fmt.Errorf("error retrieving instances: %w", err)
-	}
-
-	var instances []InstanceInfo
-	scanner := bufio.NewScanner(&buf)
-	for scanner.Scan() {
-		var i InstanceInfo
-		line := scanner.Bytes()
-		if err := json.Unmarshal(line, &i); err != nil {
-			return nil, fmt.Errorf("error retrieving instances: %w", err)
-		}
-
-		// limit to colima instances
-		if !strings.HasPrefix(i.Name, "colima") {
-			continue
-		}
-
-		// rename to local friendly names
-		i.Name = toUserFriendlyName(i.Name)
-
-		instances = append(instances, i)
-	}
-
-	return instances, nil
-}
-
-// ShowSSH runs the show-ssh command in Lima.
-func ShowSSH(name, format string) error {
-	var buf bytes.Buffer
-	cmd := cli.Command("limactl", "show-ssh", "--format", format, name)
-	cmd.Stdout = &buf
-	cmd.Stderr = os.Stderr
-
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("error retrieving ssh config: %w", err)
-	}
-
-	// TODO: this is a lazy approach, edge cases may not be covered
-	from := "lima-" + name
-	to := name
-	out := strings.ReplaceAll(buf.String(), from, to)
-
-	fmt.Println(out)
-	return nil
-}
-
-func toUserFriendlyName(name string) string {
-	if name == "colima" {
-		name = "default"
-	}
-	return strings.TrimPrefix(name, "colima-")
 }
